# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.GenerateReportRequest:
    description: 生成审计报告的请求体
    properties:
      cert_ids:
        description: 指定证书 ID 列表
        items:
          type: string
        type: array
      end_time:
        description: 结束时间
        example: "2024-12-31"
        type: string
      format:
        description: '输出格式: json/html'
        example: html
        type: string
      include_raw:
        description: 是否包含原始数据
        type: boolean
      start_time:
        description: 开始时间
        example: "2024-01-01"
        type: string
      trace_ids:
        description: 指定追踪 ID 列表
        items:
          type: string
        type: array
      type:
        description: '报告类型: audit/compliance/summary'
        example: audit
        type: string
    required:
    - type
    type: object
...
  api.VerifyCertRequest:
    description: 验证存证证书的请求体
    properties:
      cert_id:
        description: 证书 ID
        example: cert_abc123
        type: string
      root_hash:
        description: Merkle 根哈希
        example: sha256:abcd1234
        type: string
    type: object
...
  report.CertDetail:
    properties:
      cert_id:
        type: string
      created_at:
        type: string
      evidence_level:
        type: string
      root_hash:
        type: string
      trace_id:
        type: string
      verified:
        type: boolean
    type: object
  report.Details:
    properties:
      certificates:
        items:
          $ref: '#/definitions/report.CertDetail'
        type: array
      traces:
        items:
          $ref: '#/definitions/report.TraceDetail'
        type: array
      verifications:
        items:
          $ref: '#/definitions/report.Verification'
        type: array
    type: object
  report.Report:
    properties:
      content_type:
        type: string
      details:
        $ref: '#/definitions/report.Details'
      format:
        $ref: '#/definitions/report.ReportFormat'
      generated_at:
        type: string
      id:
        type: string
      summary:
        $ref: '#/definitions/report.Summary'
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/report.ReportType'
    type: object
  report.ReportFormat:
    enum:
    - json
    - html
    - pdf
    type: string
    x-enum-comments:
      ReportFormatPDF: 需要额外依赖
    x-enum-descriptions:
    - ""
    - ""
    - 需要额外依赖
    x-enum-varnames:
    - ReportFormatJSON
    - ReportFormatHTML
    - ReportFormatPDF
  report.ReportType:
    enum:
    - audit
    - compliance
    - summary
    type: string
    x-enum-comments:
      ReportTypeAudit: 审计报告
      ReportTypeCompliance: 合规报告
      ReportTypeSummary: 摘要报告
    x-enum-descriptions:
    - 审计报告
    - 合规报告
    - 摘要报告
    x-enum-varnames:
    - ReportTypeAudit
    - ReportTypeCompliance
    - ReportTypeSummary
  report.Summary:
    properties:
      certs_by_level:
        additionalProperties:
          type: integer
        type: object
      events_by_type:
        additionalProperties:
          type: integer
        type: object
      time_range:
        $ref: '#/definitions/report.TimeRange'
      token_usage:
        $ref: '#/definitions/report.TokenUsage'
      total_certificates:
        type: integer
      total_events:
        type: integer
      total_traces:
        type: integer
    type: object
  report.TimeRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
  report.TokenUsage:
    properties:
      total_completion_tokens:
        type: integer
      total_prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
  report.TraceDetail:
    properties:
      cert_id:
        type: string
      duration:
        type: string
      end_time:
        type: string
      event_count:
        type: integer
      has_cert:
        type: boolean
      start_time:
        type: string
      trace_id:
        type: string
    type: object
  report.Verification:
    properties:
      anchor_verified:
        type: boolean
      cert_id:
        type: string
      hash_integrity:
        type: boolean
      overall_valid:
        type: boolean
      signature_valid:
        type: boolean
      verified_at:
        type: string
    type: object
host: localhost:8006
info:
  contact:
    email: support@aitrace.cc
    name: AI-Trace Support
    url: https://aitrace.cc/support
  description: |-
    AI-Trace - Enterprise AI Decision Audit & Compliance Platform

    AI-Trace 提供零代码侵入的 AI 调用追踪、存证和审计能力。
    支持 OpenAI、Claude、Ollama 等多种模型，用户可透传自己的 API Key，服务不存储任何敏感信息。

    ## 特性
    - **零侵入集成**: 只需替换 API endpoint，无需修改代码
    - **API Key 透传**: 用户的 API Key 直接透传到上游，不存储
    - **自定义代理**: 支持用户指定自己的代理服务器
    - **多级存证**: internal(内部审计) / compliance(合规) / legal(法律效力)
    - **最小披露**: 基于 Merkle 树的选择性披露证明
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://aitrace.cc/terms
  title: AI-Trace API
  version: 0.2.0
paths:
...
  /certs/verify:
    post:
      consumes:
      - application/json
      description: |-
        验证存证证书的完整性和有效性

        ## 验证项目
        - **hash_integrity**: 事件哈希与 Merkle 根匹配
        - **signature**: Ed25519 数字签名验证
        - **time_proof**: 时间证明验证
        - **anchor_proof**: 锚定证明验证（WORM/区块链）
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 验证请求（cert_id 或 root_hash 二选一）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.VerifyCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 证书不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 验证存证证书
      tags:
      - Certificates
...
  /reports/generate:
    post:
      consumes:
      - application/json
      description: |-
        生成 AI 决策审计报告，支持多种类型和格式

        ## 报告类型
        - **audit**: 完整审计报告，包含所有追踪和证书详情
        - **compliance**: 合规性报告，重点关注证书验证状态
        - **summary**: 摘要报告，统计汇总信息

        ## 输出格式
        - **json**: JSON 格式，适合程序处理
        - **html**: HTML 格式，适合直接查看和打印
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 报告请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GenerateReportRequest'
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: 报告信息
          schema:
            $ref: '#/definitions/report.Report'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 生成审计报告
      tags:
      - Reports
securityDefinitions:
  ApiKeyAuth:
    description: AI-Trace 平台 API Key（用于认证）
    in: header
    name: X-API-Key
    type: apiKey
  UpstreamApiKey:
    description: 用户的上游 API Key（OpenAI/Claude，透传不存储）
    in: header
    name: X-Upstream-API-Key
    type: apiKey
swagger: "2.0"